Time: 1 ms (99.47%), Space: 42 MB (20.42%) - LeetHub

+++ 3110-score-of-a-string/3110-score-of-a-string.java
@@ -4,7 +4,7 @@ class Solution {
         int ans = 0;
         
         for(int i=1;i<s.length();++i)
-                ans += Math.abs((int)s.charAt(i)- (int)s.charAt(i-1));
+                ans += Math.abs(s.charAt(i) - s.charAt(i-1));
         return ans;
     }
 }
